requirements.txt
flask
python-dotenv
aiohttp
requests
colorama
flask-cors
flask-migrate